+++ crates/kitsune_p2p/kitsune_p2p/Cargo.toml
@@ -17,7 +17,6 @@ ghost_actor = "0.2.1"
 kitsune_p2p_types = { version = "0.0.1", path = "../types" }
 kitsune_p2p_proxy = { version = "0.0.1", path = "../proxy" }
 kitsune_p2p_transport_quic = { version = "0.0.1", path = "../transport_quic" }
-once_cell = "1"
 rand = "0.7"
 shrinkwraprs = "0.3.0"
 thiserror = "1.0.18"
